fix min-triggers to  1
The purpose of this pull request is to enable simulation when only one detector (H1 or L1 or V1 or K1) is in operation during a run. There is no coincidence . This ovoid the error. For more https://github.com/lpsinger/ligo.skymap/blob/main/ligo/skymap/tool/bayestar_inject.py#L318